refactor modules to remove regex from tests
