+++ src/asynch/transaction/mod.rs
@@ -4,6 +4,10 @@ mod submit_and_wait;
 use bigdecimal::{BigDecimal, RoundingMode};
 pub use submit_and_wait::*;
 
+// `sign` lives in `crate::signing` (pure crypto, no async client required).
+// Re-exported here for backward compatibility.
+pub use crate::signing::sign;
+
 use crate::{
     asynch::{
         account::get_next_valid_seq_number,
@@ -11,21 +15,12 @@ use crate::{
         ledger::{get_fee, get_latest_validated_ledger_sequence},
         transaction::exceptions::XRPLSignTransactionException,
     },
-    core::{
-        addresscodec::{is_valid_xaddress, xaddress_to_classic_address},
-        binarycodec::{encode, encode_for_multisigning, encode_for_signing},
-        keypairs::sign as keypairs_sign,
-    },
+    core::binarycodec::encode,
     models::{
         requests::{server_state::ServerState, submit::Submit},
         results::{server_state::ServerState as ServerStateResult, submit::Submit as SubmitResult},
-        transactions::{
-            exceptions::XRPLTransactionFieldException, Signer, Transaction, TransactionType,
-        },
+        transactions::{Transaction, TransactionType},
         Model, XRPAmount, XRPLModelException,
-    },
-    utils::transactions::{
-        get_transaction_field_value, set_transaction_field_value, validate_transaction_has_field,
     },
     wallet::Wallet,
 };
@@ -33,12 +28,12 @@ use crate::{
 use alloc::string::String;
 use alloc::string::ToString;
 use alloc::vec::Vec;
-use alloc::{borrow::Cow, vec};
+use alloc::borrow::Cow;
 use core::convert::TryInto;
 use core::fmt::Debug;
 use exceptions::XRPLTransactionHelperException;
 use serde::Serialize;
-use serde::{de::DeserializeOwned, Deserialize};
+use serde::de::DeserializeOwned;
 use strum::IntoEnumIterator;
 
 use super::exceptions::XRPLHelperResult;
@@ -47,40 +42,6 @@ const OWNER_RESERVE: &str = "2000000"; // 2 XRP
 const RESTRICTED_NETWORKS: u16 = 1024;
 const REQUIRED_NETWORKID_VERSION: &str = "1.11.0";
 const LEDGER_OFFSET: u8 = 20;
-
-pub fn sign<'a, T, F>(transaction: &mut T, wallet: &Wallet, multisign: bool) -> XRPLHelperResult<()>
-where
-    F: IntoEnumIterator + Serialize + Debug + PartialEq,
-    T: Transaction<'a, F> + Model + Serialize + DeserializeOwned + Clone + Debug,
-{
-    transaction.validate()?;
-
-    if multisign {
-        let serialized_for_signing =
-            encode_for_multisigning(transaction, wallet.classic_address.clone().into())?;
-        let serialized_bytes = hex::decode(serialized_for_signing)?;
-        let signature = keypairs_sign(&serialized_bytes, &wallet.private_key)?;
-        let signer = Signer::new(
-            wallet.classic_address.clone(),
-            signature,
-            wallet.public_key.clone(),
-        );
-        transaction.get_mut_common_fields().signers = Some(vec![signer]);
-
-        Ok(())
-    } else {
-        prepare_transaction(transaction, wallet)?;
-        // dbg!("Transaction prepared");
-        let serialized_for_signing = encode_for_signing(transaction)?;
-        let serialized_bytes = hex::decode(serialized_for_signing)?;
-        // dbg!("Transaction serialized");
-        let signature = keypairs_sign(&serialized_bytes, &wallet.private_key)?;
-        // dbg!("Signature created");
-        transaction.get_mut_common_fields().txn_signature = Some(signature.into());
-
-        Ok(())
-    }
-}
 
 pub async fn sign_and_submit<'a, 'b, T, F, C>(
     transaction: &mut T,
@@ -354,12 +315,6 @@ fn is_not_later_rippled_version(source: String, target: String) -> XRPLHelperRes
     }
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
-enum AccountFieldType {
-    Account,
-    Destination,
-}
-
 async fn check_txn_fee<'a, 'b, T, F, C>(transaction: &mut T, client: &'b C) -> XRPLHelperResult<()>
 where
     F: IntoEnumIterator + Serialize + Debug + PartialEq,
@@ -376,105 +331,6 @@ where
         .unwrap_or(XRPAmount::from("0"));
     if transaction_fee > expected_fee {
         Err(XRPLSignTransactionException::FeeTooHigh(transaction_fee.to_string()).into())
-    } else {
-        Ok(())
-    }
-}
-
-fn prepare_transaction<'a, T, F>(transaction: &mut T, wallet: &Wallet) -> XRPLHelperResult<()>
-where
-    F: IntoEnumIterator + Serialize + Debug + PartialEq,
-    T: Transaction<'a, F> + Serialize + DeserializeOwned + Clone,
-{
-    let commond_fields = transaction.get_mut_common_fields();
-    commond_fields.signing_pub_key = Some(wallet.public_key.clone().into());
-
-    validate_account_xaddress(transaction, AccountFieldType::Account)?;
-    if validate_transaction_has_field(transaction, "Destination").is_ok() {
-        validate_account_xaddress(transaction, AccountFieldType::Destination)?;
-    }
-
-    let _ = convert_to_classic_address(transaction, "Unauthorize");
-    let _ = convert_to_classic_address(transaction, "Authorize");
-    // EscrowCancel, EscrowFinish
-    let _ = convert_to_classic_address(transaction, "Owner");
-    // SetRegularKey
-
-    let _ = convert_to_classic_address(transaction, "RegularKey");
-
-    Ok(())
-}
-
-fn validate_account_xaddress<'a, T, F>(
-    prepared_transaction: &mut T,
-    account_field: AccountFieldType,
-) -> XRPLHelperResult<()>
-where
-    F: IntoEnumIterator + Serialize + Debug + PartialEq,
-    T: Transaction<'a, F> + Serialize + DeserializeOwned + Clone,
-{
-    let (account_field_name, tag_field_name) = match serde_json::to_string(&account_field) {
-        Ok(name) => {
-            let name_str = name.as_str().trim();
-            if name_str == "\"Account\"" {
-                ("Account", "SourceTag")
-            } else if name_str == "\"Destination\"" {
-                ("Destination", "DestinationTag")
-            } else {
-                return Err(XRPLTransactionFieldException::UnknownAccountField(
-                    name_str.to_string(),
-                )
-                .into());
-            }
-        }
-        Err(error) => return Err(error.into()),
-    };
-    let account_address = match account_field {
-        AccountFieldType::Account => prepared_transaction.get_common_fields().account.clone(),
-        AccountFieldType::Destination => {
-            get_transaction_field_value(prepared_transaction, "Destination")?
-        }
-    };
-
-    if is_valid_xaddress(&account_address) {
-        let (address, tag, _) = xaddress_to_classic_address(&account_address)?;
-        validate_transaction_has_field(prepared_transaction, account_field_name)?;
-        set_transaction_field_value(prepared_transaction, account_field_name, address)?;
-
-        if validate_transaction_has_field(prepared_transaction, tag_field_name).is_ok()
-            && get_transaction_field_value(prepared_transaction, tag_field_name).unwrap_or(Some(0))
-                != tag
-        {
-            Err(XRPLSignTransactionException::TagFieldMismatch(tag_field_name.to_string()).into())
-        } else {
-            set_transaction_field_value(prepared_transaction, tag_field_name, tag)?;
-
-            Ok(())
-        }
-    } else {
-        Ok(())
-    }
-}
-
-fn convert_to_classic_address<'a, T, F>(
-    transaction: &mut T,
-    field_name: &str,
-) -> XRPLHelperResult<()>
-where
-    F: IntoEnumIterator + Serialize + Debug + PartialEq,
-    T: Transaction<'a, F> + Serialize + DeserializeOwned + Clone,
-{
-    let address = get_transaction_field_value::<F, _, String>(transaction, field_name)?;
-    if is_valid_xaddress(&address) {
-        let classic_address = match xaddress_to_classic_address(&address) {
-            Ok(t) => t.0,
-            Err(error) => return Err(error.into()),
-        };
-        Ok(set_transaction_field_value(
-            transaction,
-            field_name,
-            classic_address,
-        )?)
     } else {
         Ok(())
     }
